pyproject.toml
[tool.poetry]
name = "cats_and_dogs"
version = "0.1.0"
description = "MLOps example"
authors = ["Vladislav Naumov <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fire = "^0.7.0"
torch = "^2.5.0"
torchvision = "^0.20.1"
pytorch-lightning = "^2.4.0"
pandas="^2.0.0"
requests="*"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
sphinx = "^8.1.3"
dvc = "^3.58.0"
wandb = "^0.19.1"
mlflow = "^2.19.0"

[tool.isort]
src_paths = ["cats_and_dogs"]
profile = "black"
line_length = 90
lines_after_imports = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
